Moved AcqPosSetpoints to utils folder
