+++ src/main/java/frc/robot/OperatorState.java
@@ -1,5 +1,7 @@
 package frc.robot;
 
+import frc.robot.utils.AcquisitionPositionSetpoint;
+
 public class OperatorState {
     private AcquisitionPositionSetpoint lastInput;
     private ScoringSide scoringSide;
